src/main/webapp/app/entities/instituicao-ensino/delete/instituicao-ensino-delete-dialog.component.ts
import { Component } from '@angular/core';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { IInstituicaoEnsino } from '../instituicao-ensino.model';
import { InstituicaoEnsinoService } from '../service/instituicao-ensino.service';
import { ITEM_DELETED_EVENT } from 'app/config/navigation.constants';

@Component({
  templateUrl: './instituicao-ensino-delete-dialog.component.html',
})
export class InstituicaoEnsinoDeleteDialogComponent {
  instituicaoEnsino?: IInstituicaoEnsino;

  constructor(protected instituicaoEnsinoService: InstituicaoEnsinoService, protected activeModal: NgbActiveModal) {}

  cancel(): void {
    this.activeModal.dismiss();
  }

  confirmDelete(id: number): void {
    this.instituicaoEnsinoService.delete(id).subscribe(() => {
      this.activeModal.close(ITEM_DELETED_EVENT);
    });
  }
}
